requirements.txt
postgres
django
djangorestframework
django-filter
djangorestframework-jwt
djangorestframework-recursive
